Clear out more code.
