+++ cpp/dolfin/mesh/CellType.h
@@ -40,9 +40,6 @@ public:
   virtual double squared_distance(const Cell& cell,
                                   const Eigen::Vector3d& point) const = 0;
 
-  /// Compute of given facet with respect to the cell
-  virtual Eigen::Vector3d normal(const Cell& cell, std::size_t facet) const = 0;
-
   const CellType type;
 };
 } // namespace mesh
