+++ cpp/dolfin/mesh/QuadrilateralCell.h
@@ -26,9 +26,6 @@ public:
   /// Compute squared distance to given point (3D enabled)
   double squared_distance(const mesh::Cell& cell,
                           const Eigen::Vector3d& point) const;
-
-  /// Compute of given facet with respect to the cell
-  Eigen::Vector3d normal(const mesh::Cell& cell, std::size_t facet) const;
 };
 } // namespace mesh
 } // namespace dolfin
