+++ cpp/dolfin/mesh/TetrahedronCell.h
@@ -27,10 +27,6 @@ public:
   double squared_distance(const mesh::Cell& cell,
                           const Eigen::Vector3d& point) const;
 
-  /// Compute component normal of given facet with respect to
-  /// the cell
-  Eigen::Vector3d normal(const mesh::Cell& cell, std::size_t facet) const;
-
 private:
   // Find local index of edge i according to ordering convention
   std::size_t find_edge(std::size_t i, const mesh::Cell& cell) const;
