+++ cpp/dolfin/mesh/TriangleCell.h
@@ -35,9 +35,6 @@ public:
                                  const Eigen::Vector3d& b,
                                  const Eigen::Vector3d& c);
 
-  /// Compute of given facet with respect to the cell
-  Eigen::Vector3d normal(const mesh::Cell& cell, std::size_t facet) const;
-
 private:
   // Find local index of edge i according to ordering convention
   std::size_t find_edge(std::size_t i, const mesh::Cell& cell) const;
