+++ cpp/dolfin/mesh/utils.cpp
@@ -6,6 +6,7 @@
 
 #include "utils.h"
 #include "Cell.h"
+#include "Facet.h"
 #include "Geometry.h"
 #include "MeshEntity.h"
 #include <Eigen/Dense>
@@ -885,6 +886,133 @@ Eigen::Vector3d mesh::cell_normal(const mesh::Cell& cell)
     throw std::invalid_argument(
         "cell_normal not supported for this cell type.");
   }
+  return Eigen::Vector3d();
+}
+//-----------------------------------------------------------------------------
+Eigen::Vector3d mesh::normal(const mesh::Cell& cell, int facet)
+{
+  const mesh::Geometry& geometry = cell.mesh().geometry();
+  const mesh::CellType type = cell.type();
+
+  switch (type)
+  {
+  case (mesh::CellType::interval):
+  {
+    const std::int32_t* vertices = cell.entities(0);
+    Eigen::Vector3d n = geometry.x(vertices[0]) - geometry.x(vertices[1]);
+    n.normalize();
+    if (facet == 1)
+      return -1.0 * n;
+    else
+      return n;
+  }
+  case (mesh::CellType::triangle):
+  {
+    // The normal vector is currently only defined for a triangle in R^2
+    // MER: This code is super for a triangle in R^3 too, this error
+    // could be removed, unless it is here for some other reason.
+    if (geometry.dim() != 2)
+      throw std::runtime_error("Illegal geometric dimension");
+
+    cell.mesh().create_connectivity(2, 1);
+    mesh::Facet f(cell.mesh(), cell.entities(1)[facet]);
+
+    // Get global index of opposite vertex
+    const int v0 = cell.entities(0)[facet];
+
+    // Get global index of vertices on the facet
+    const int v1 = f.entities(0)[0];
+    const int v2 = f.entities(0)[1];
+
+    // Get the coordinates of the three vertices
+    const Eigen::Vector3d p0 = geometry.x(v0);
+    const Eigen::Vector3d p1 = geometry.x(v1);
+    const Eigen::Vector3d p2 = geometry.x(v2);
+
+    // Subtract projection of p2 - p0 onto p2 - p1
+    Eigen::Vector3d t = p2 - p1;
+    Eigen::Vector3d n = p2 - p0;
+    t.normalize();
+    n.normalize();
+    n -= t * n.dot(t);
+    n.normalize();
+    return n;
+  }
+  case (mesh::CellType::quadrilateral):
+  {
+    // Make sure we have facets
+    cell.mesh().create_connectivity(2, 1);
+
+    // Create facet from the mesh and local facet number
+    Facet f(cell.mesh(), cell.entities(1)[facet]);
+
+    if (cell.mesh().geometry().dim() != 2)
+      throw std::runtime_error("Illegal geometric dimension");
+
+    // Get global index of opposite vertex
+    const std::size_t v0 = cell.entities(0)[facet];
+
+    // Get global index of vertices on the facet
+    const std::size_t v1 = f.entities(0)[0];
+    const std::size_t v2 = f.entities(0)[1];
+
+    // Get the coordinates of the three vertices
+    const Eigen::Vector3d p0 = geometry.x(v0);
+    const Eigen::Vector3d p1 = geometry.x(v1);
+    const Eigen::Vector3d p2 = geometry.x(v2);
+
+    // Subtract projection of p2 - p0 onto p2 - p1
+    Eigen::Vector3d t = p2 - p1;
+    t.normalize();
+    Eigen::Vector3d n = p2 - p0;
+    n -= t * n.dot(t);
+    n.normalize();
+    return n;
+  }
+  case (mesh::CellType::tetrahedron):
+  {
+    // Make sure we have facets
+    cell.mesh().create_connectivity(3, 2);
+
+    // Create facet from the mesh and local facet number
+    Facet f(cell.mesh(), cell.entities(2)[facet]);
+
+    // Get global index of opposite vertex
+    const std::size_t v0 = cell.entities(0)[facet];
+
+    // Get global index of vertices on the facet
+    std::size_t v1 = f.entities(0)[0];
+    std::size_t v2 = f.entities(0)[1];
+    std::size_t v3 = f.entities(0)[2];
+
+    // Get the coordinates of the four vertices
+    const Eigen::Vector3d P0 = geometry.x(v0);
+    const Eigen::Vector3d P1 = geometry.x(v1);
+    const Eigen::Vector3d P2 = geometry.x(v2);
+    const Eigen::Vector3d P3 = geometry.x(v3);
+
+    // Create vectors
+    Eigen::Vector3d V0 = P0 - P1;
+    Eigen::Vector3d V1 = P2 - P1;
+    Eigen::Vector3d V2 = P3 - P1;
+
+    // Compute normal vector
+    Eigen::Vector3d n = V1.cross(V2);
+
+    // Normalize
+    n.normalize();
+
+    // Flip direction of normal so it points outward
+    if (n.dot(V0) > 0)
+      n *= -1.0;
+
+    return n;
+  }
+  // case (mesh::CellType::hexahedron):
+  default:
+    throw std::runtime_error("Unknown cell type.");
+  }
+
   return Eigen::Vector3d();
 }
 //-----------------------------------------------------------------------------
